Create Challenge with Tasks

+++ Habitica/src/main/java/com/habitrpg/android/habitica/ui/activities/MainActivity.java
@@ -1120,6 +1120,10 @@ public class MainActivity extends BaseActivity implements Action1<Throwable>, Ha
 
     @Subscribe
     public void onEvent(final DeleteTaskCommand cmd) {
+        if(cmd.ignoreEvent) {
+            return;
+        }
+
         apiClient.deleteTask(cmd.TaskIdToDelete)
                 .subscribe(aVoid -> {
                     EventBus.getDefault().post(new TaskRemovedEvent(cmd.TaskIdToDelete));
@@ -1464,6 +1468,9 @@ public class MainActivity extends BaseActivity implements Action1<Throwable>, Ha
 
     @Subscribe
     public void onEvent(final TaskSaveEvent event) {
+        if(event.ignoreEvent)
+            return;
+
         Task task = event.task;
         if (event.created) {
             this.taskRepository.createTask(task).subscribe(new TaskCreationCallback(), throwable -> {
